MasterMind Final Commit
MasterMind is a two player game where you'll set a  4 digit code for your opponent and he'll do same for you. You have to guess the opponent's code which is set by him. You have to enter 4 distinct digits and you'll be given clue of how many digits are in correct position and in wrong position in your guessed code.
Same for opponent also. He'll be given the same clue and has to guess your code.